refactor(mm-next/story): add shared component `hero-image-and-video` in premium and wide story
